requirements.txt
anthropic
sentence-transformers
faiss-cpu
python-dotenv